전체적으로 구현완료.. 단위테스트 필요

+++ mustard/src/main/java/com/spring/domain/ZipVO.java
@@ -4,6 +4,6 @@ import lombok.Data;
  
 @Data
 public class ZipVO {
-	private int code;
+	private long code;
 	private String shi, gungu, dong;
 }
